Move panel logic to dedicated handler
