+++ src/platform_layer/command_executor.rs
@@ -521,131 +521,6 @@ pub(crate) fn execute_close_window(
     super::window_common::send_close_message(internal_state, window_id)
 }
 
-/*
- * Custom window procedure for panel controls.
- * It forwards important messages like WM_COMMAND and WM_CTLCOLOR* to the panel's parent
- * so that controls embedded within the panel generate events and can be custom-drawn
- * at the main window level.
- */
-unsafe extern "system" fn forwarding_panel_proc(
-    hwnd: HWND,
-    msg: u32,
-    wparam: WPARAM,
-    lparam: LPARAM,
-) -> LRESULT {
-    unsafe {
-        // Forward key messages to the parent window (the main window).
-        // This allows the main window's WndProc to handle notifications from controls inside the panel.
-        if msg == WM_COMMAND
-            || msg == WM_CTLCOLOREDIT
-            || msg == WM_CTLCOLORSTATIC
-            || msg == WM_NOTIFY
-        {
-            if let Ok(parent) = GetParent(hwnd) {
-                if !parent.is_invalid() {
-                    // Use SendMessageW to synchronously send the message and return the result.
-                    // This is crucial for messages like WM_CTLCOLOREDIT that expect a return value (HBRUSH).
-                    return SendMessageW(parent, msg, Some(wparam), Some(lparam));
-                }
-            }
-        }
-
-        let prev = GetWindowLongPtrW(hwnd, GWLP_USERDATA);
-        if prev != 0 {
-            let prev_proc: WNDPROC = std::mem::transmute(prev);
-            return CallWindowProcW(prev_proc, hwnd, msg, wparam, lparam);
-        }
-        DefWindowProcW(hwnd, msg, wparam, lparam)
-    }
-}
-
-/*
- * Executes the `CreatePanel` command.
- * Creates a generic STATIC control to act as a panel. The panel can be a child
- * of the main window or another control (identified by `parent_control_id`).
- * The new panel's HWND is stored in `NativeWindowData.controls` mapped by `panel_id`.
- */
-pub(crate) fn execute_create_panel(
-    internal_state: &Arc<Win32ApiInternalState>,
-    window_id: WindowId,
-    parent_control_id: Option<i32>,
-    panel_id: i32, // Logical ID for this new panel
-) -> PlatformResult<()> {
-    log::debug!(
-        "CommandExecutor: execute_create_panel for WinID {:?}, PanelID: {}, ParentControlID: {:?}",
-        window_id,
-        panel_id,
-        parent_control_id
-    );
-
-    internal_state.with_window_data_write(window_id, |window_data| {
-        if window_data.has_control(panel_id) {
-            log::warn!(
-                "CommandExecutor: Panel with logical ID {} already exists for window {:?}.",
-                panel_id,
-                window_id
-            );
-            return Err(PlatformError::OperationFailed(format!(
-                "Panel with logical ID {} already exists for window {:?}",
-                panel_id, window_id
-            )));
-        }
-
-        let hwnd_parent = match parent_control_id {
-            Some(id) => window_data.get_control_hwnd(id).ok_or_else(|| {
-                log::warn!("CommandExecutor: Parent control with logical ID {} not found for CreatePanel in WinID {:?}", id, window_id);
-                PlatformError::InvalidHandle(format!(
-                    "Parent control with logical ID {} not found for CreatePanel in WinID {:?}",
-                    id, window_id
-                ))
-            })?,
-            None => window_data.get_hwnd(), // Parent is the main window
-        };
-
-        if hwnd_parent.is_invalid() {
-            log::error!(
-                "CommandExecutor: Parent HWND for CreatePanel is invalid (WinID: {:?}, ParentControlID: {:?})",
-                window_id,
-                parent_control_id
-            );
-            return Err(PlatformError::InvalidHandle(format!(
-                "Parent HWND for CreatePanel is invalid (WinID: {:?}, ParentControlID: {:?})",
-                window_id, parent_control_id
-            )));
-        }
-
-        let h_instance = internal_state.h_instance();
-        let hwnd_panel = unsafe {
-            CreateWindowExW(
-                WINDOW_EX_STYLE(0), // Or WS_EX_CONTROLPARENT if it should manage tab order for children
-                super::window_common::WC_STATIC, // Using a STATIC control as a simple panel container
-                None,               // No text for a simple panel
-                WS_CHILD | WS_VISIBLE, // Basic styles for a panel
-                0,
-                0,
-                10,
-                10, // Dummy position/size, layout rules will adjust
-                Some(hwnd_parent),
-                Some(HMENU(panel_id as *mut _)), // Use logical ID for the HMENU
-                Some(h_instance),
-                None,
-            )?
-        };
-        unsafe {
-            #[allow(clippy::fn_to_numeric_cast)]
-            let prev = SetWindowLongPtrW(hwnd_panel, GWLP_WNDPROC, forwarding_panel_proc as isize);
-            SetWindowLongPtrW(hwnd_panel, GWLP_USERDATA, prev);
-        }
-        window_data.register_control_hwnd(panel_id, hwnd_panel);
-        log::debug!(
-            "CommandExecutor: Created panel (LogicalID {}) for WinID {:?} with HWND {:?}",
-            panel_id,
-            window_id,
-            hwnd_panel
-        );
-        Ok(())
-    })
-}
 
 #[cfg(test)]
 mod tests {
